Extract game input state handling from IngameState
Prevents input getting "stuck" when input is removed in the pause menu
