+++ rwgame/RWGame.cpp
@@ -1,5 +1,6 @@
 #include "RWGame.hpp"
 #include "DrawUI.hpp"
+#include "GameInput.hpp"
 #include "State.hpp"
 #include "states/BenchmarkState.hpp"
 #include "states/IngameState.hpp"
@@ -393,6 +394,8 @@ int RWGame::run() {
                     event.motion.yrel *= MOUSE_SENSITIVITY_SCALE;
                     break;
             }
+
+            GameInput::updateGameInputState(&getState()->input[0], event);
 
             RW_PROFILE_BEGIN("State");
             state->handleEvent(event);
