+++ rwgame/states/IngameState.hpp
@@ -53,7 +53,6 @@ public:
 
     virtual void handleEvent(const SDL_Event& event);
     virtual void handlePlayerInput(const SDL_Event& event);
-    void updateInputState(const SDL_Event& event);
 
     virtual bool shouldWorldUpdate();
